requirements.txt
streamlit
pandas
numpy
matplotlib
Seaborn
scipy
scikit-learn
plotly
google-generativeai
requests
tabulate